chore: factor out is_unique from duckdb and pyspark
