+++ narwhals/_ibis/expr.py
@@ -396,7 +396,7 @@ class IbisExpr(SQLExpr["IbisLazyFrame", "ir.Value"]):
     skew = not_implemented()
     kurtosis = not_implemented()
 
-    _star = not_implemented()
+    _count_star = not_implemented()
 
     # Intentionally not implemented, as Ibis does its own expression rewriting.
     _push_down_window_function = not_implemented()
